schaaaafrichter/requirements.txt
chainer==4.0.0
chainercv==0.9.0
cupy==4.0.0
numpy==1.14.2
opencv-python>=3.4.0.12
Pillow==5.1.0
simpleaudio==1.0.2
tqdm==4.23.1
